--- a/src/query/sql/src/planner/optimizer/distributed/distributed.rs
+++ b/src/query/sql/src/planner/optimizer/distributed/distributed.rs
@@ -17,18 +17,21 @@ use std::sync::Arc;
 use common_catalog::table_context::TableContext;
 use common_exception::Result;
 
+use crate::optimizer::distributed::topk::TopK;
 use crate::optimizer::property::require_property;
 use crate::optimizer::Distribution;
 use crate::optimizer::RelExpr;
 use crate::optimizer::RequiredProperty;
 use crate::optimizer::SExpr;
 use crate::plans::Exchange;
+use crate::plans::RelOperator;
 
 pub fn optimize_distributed_query(ctx: Arc<dyn TableContext>, s_expr: &SExpr) -> Result<SExpr> {
     let required = RequiredProperty {
         distribution: Distribution::Any,
     };
     let mut result = require_property(ctx, &required, s_expr)?;
+    push_down_topk_to_merge(&mut result, None)?;
     let rel_expr = RelExpr::with_s_expr(&result);
     let physical_prop = rel_expr.derive_physical_prop()?;
     let root_required = RequiredProperty {
@@ -40,4 +43,28 @@ pub fn optimize_distributed_query(ctx: Arc<dyn TableContext>, s_expr: &SExpr) ->
     }
 
     Ok(result)
+}
+
+// Traverse the SExpr tree to find top_k, if find, push down it to Exchange::Merge
+fn push_down_topk_to_merge(s_expr: &mut SExpr, mut top_k: Option<TopK>) -> Result<()> {
+    if let RelOperator::Exchange(Exchange::Merge) = s_expr.plan {
+        if let Some(top_k) = top_k {
+            let child = &mut s_expr.children[0];
+            *child = SExpr::create_unary(top_k.sort.into(), child.clone());
+            *child = SExpr::create_unary(top_k.limit.into(), child.clone());
+        }
+        return Ok(());
+    }
+    for child in s_expr.children.iter_mut() {
+        if let RelOperator::Sort(sort) = &child.plan {
+            if let RelOperator::Limit(limit) = &s_expr.plan {
+                top_k = Some(TopK {
+                    sort: sort.clone(),
+                    limit: limit.clone(),
+                });
+            }
+        }
+        push_down_topk_to_merge(child, top_k.clone())?;
+    }
+    Ok(())
 }