I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

The ticket pushes down `Limit + Order by` to each node in cluster.

I did a simple performance and there is **~40x** performance improvement, I think tpch benchmark also will leverage it.

```sql
Before

mysql> SELECT number FROM numbers_mt(10000000000) ORDER BY number DESC LIMIT 10;
+------------+
| number     |
+------------+
| 9999999999 |
| 9999999998 |
| 9999999997 |
| 9999999996 |
| 9999999995 |
| 9999999994 |
| 9999999993 |
| 9999999992 |
| 9999999991 |
| 9999999990 |
+------------+
10 rows in set (1 min 7.91 sec)
Read 10000000000 rows, 74.51 GiB in 67.910 sec., 147.25 million rows/sec., 1.10 GiB/sec.
Now
mysql> SELECT number FROM numbers_mt(10000000000) ORDER BY number DESC LIMIT 10;
+------------+
| number     |
+------------+
| 9999999999 |
| 9999999998 |
| 9999999997 |
| 9999999996 |
| 9999999995 |
| 9999999994 |
| 9999999993 |
| 9999999992 |
| 9999999991 |
| 9999999990 |
+------------+
10 rows in set (1.77 sec)
Read 10000000000 rows, 74.51 GiB in 1.763 sec., 5.67 billion rows/sec., 42.27 GiB/sec.
```
Closes https://github.com/datafuselabs/databend/issues/6525 https://github.com/datafuselabs/databend/issues/7015